refactor: normalize backend state and deploy names
Consolidate backend deploy/state naming, remove stale deploy surfaces, and keep the validated route-authority cleanup documented.\n\nCloses #126
