+++ crates/local_backend/src/router.rs
@@ -81,11 +81,9 @@ use crate::{
         run_test_function,
     },
     deploy_config::{
-        get_config,
+        self,
         get_config_hashes,
-        push_config,
     },
-    deploy_config2,
     environment_variables::{
         list_environment_variables,
         platform_router,
@@ -123,10 +121,6 @@ use crate::{
     scheduling::{
         cancel_all_jobs,
         cancel_job,
-    },
-    schema::{
-        prepare_schema,
-        schema_state,
     },
     snapshot_export::{
         cancel_export,
@@ -170,9 +164,9 @@ use crate::{
     subs::sync,
     ActionCallbackRouterState,
     AdminRouterState,
+    BackendAppState,
     DeployRouterState,
     ImportExportRouterState,
-    LocalAppState,
     RouterState,
 };
 
@@ -364,7 +358,7 @@ struct PublicApiDoc;
 )]
 struct DashboardApiDoc;
 
-pub fn router(st: LocalAppState) -> Router {
+pub fn router(st: BackendAppState) -> Router {
     let browser_routes = Router::new()
         // Called by the browser (and optionally authenticated by a cookie or `Authorization`
         // header). Passes version in the URL because websockets can't do it in header.
@@ -405,22 +399,17 @@ pub fn router(st: LocalAppState) -> Router {
         .with_state(admin_state.clone());
 
     let deploy_push_routes = Router::new()
-        .route("/push_config", post(push_config))
-        .route("/prepare_schema", post(prepare_schema))
-        .route("/deploy2/start_push", post(deploy_config2::start_push))
-        .route(
-            "/deploy2/evaluate_push",
-            post(deploy_config2::evaluate_push),
-        )
+        .route("/deploy/start_push", post(deploy_config::start_push))
+        .route("/deploy/evaluate_push", post(deploy_config::evaluate_push))
         .route("/run_test_function", post(run_test_function))
         .route(
-            "/deploy2/wait_for_schema",
-            post(deploy_config2::wait_for_schema),
+            "/deploy/wait_for_schema",
+            post(deploy_config::wait_for_schema),
         )
-        .route("/deploy2/finish_push", post(deploy_config2::finish_push))
+        .route("/deploy/finish_push", post(deploy_config::finish_push))
         .route(
-            "/deploy2/report_push_completed",
-            post(deploy_config2::report_push_completed_handler),
+            "/deploy/report_push_completed",
+            post(deploy_config::report_push_completed_handler),
         )
         .layer(
             ServiceBuilder::new()
@@ -433,9 +422,7 @@ pub fn router(st: LocalAppState) -> Router {
     let deploy_state = DeployRouterState::from(&st);
     let deploy_routes = Router::new()
         .merge(deploy_push_routes)
-        .route("/get_config", post(get_config))
         .route("/get_config_hashes", post(get_config_hashes))
-        .route("/schema_state/{schema_id}", get(schema_state))
         .layer(cli_cors())
         .layer(axum::middleware::from_fn_with_state(
             deploy_state.clone(),
@@ -688,13 +675,13 @@ where
 
 pub fn health_check_routes<S>(version: String) -> Router<S>
 where
-    LocalAppState: FromMtState<S>,
+    BackendAppState: FromMtState<S>,
     S: Clone + Send + Sync + 'static,
 {
     Router::new()
         .route(
             "/instance_name",
-            get(|MtState(st): MtState<LocalAppState>| async move { st.instance_name.clone() }),
+            get(|MtState(st): MtState<BackendAppState>| async move { st.instance_name.clone() }),
         )
         .route("/instance_version", get(|| async move { version }))
         .route(
@@ -888,37 +875,34 @@ mod tests {
     }
 
     #[convex_macro::prod_rt_test]
-    async fn test_deploy_api_rejects_non_authority_partition(
-        rt: ProdRuntime,
-    ) -> anyhow::Result<()> {
+    async fn test_legacy_deploy_routes_are_not_registered(rt: ProdRuntime) -> anyhow::Result<()> {
         let backend = setup_backend_for_test(rt).await?;
-        let mut partitioned_state = backend.st.clone();
-        partitioned_state.partition_id = Some(PartitionId(1));
-        let partitioned_app = ConvexHttpService::new(
-            router(partitioned_state),
-            "deploy_api_authority_test",
+        let app = ConvexHttpService::new(
+            router(backend.st.clone()),
+            "legacy_deploy_routes_test",
             SERVER_VERSION_STR.to_string(),
             MAX_CONCURRENT_REQUESTS,
             Duration::from_secs(125),
             NoopRouteMapper,
         );
 
-        let req = Request::builder()
-            .uri("/api/push_config")
-            .method("POST")
-            .header("Host", "localhost")
-            .header("Content-Type", "application/json")
-            .body(Body::from("{}"))?;
-        let (parts, body) = partitioned_app
-            .router()
-            .clone()
-            .oneshot(req)
-            .await?
-            .into_parts();
-        let bytes = body.collect().await?.to_bytes();
-        let body = String::from_utf8_lossy(&bytes);
-        assert_eq!(parts.status, StatusCode::SERVICE_UNAVAILABLE, "{body}");
-        assert!(body.contains("ServiceUnavailable"), "{body}");
+        for (method, path) in [
+            ("POST", "/api/get_config"),
+            ("POST", "/api/push_config"),
+            ("POST", "/api/prepare_schema"),
+            ("GET", "/api/schema_state/not-a-real-schema-id"),
+        ] {
+            let req = Request::builder()
+                .uri(path)
+                .method(method)
+                .header("Host", "localhost")
+                .header("Content-Type", "application/json")
+                .body(Body::from("{}"))?;
+            let (parts, body) = app.router().clone().oneshot(req).await?.into_parts();
+            let bytes = body.collect().await?.to_bytes();
+            let body = String::from_utf8_lossy(&bytes);
+            assert_eq!(parts.status, StatusCode::NOT_FOUND, "{path}: {body}");
+        }
 
         Ok(())
     }
@@ -963,44 +947,6 @@ mod tests {
         let bytes = body.collect().await?.to_bytes();
         let body = String::from_utf8_lossy(&bytes);
         assert_eq!(parts.status, StatusCode::OK, "{body}");
-
-        Ok(())
-    }
-
-    #[convex_macro::prod_rt_test]
-    async fn test_schema_state_rejects_non_authority_partition(
-        rt: ProdRuntime,
-    ) -> anyhow::Result<()> {
-        let backend = setup_backend_for_test(rt).await?;
-        let admin_header = backend.admin_auth_header.0.encode();
-
-        let mut partitioned_state = backend.st.clone();
-        partitioned_state.partition_id = Some(PartitionId(1));
-        let partitioned_app = ConvexHttpService::new(
-            router(partitioned_state),
-            "schema_state_authority_test",
-            SERVER_VERSION_STR.to_string(),
-            MAX_CONCURRENT_REQUESTS,
-            Duration::from_secs(125),
-            NoopRouteMapper,
-        );
-
-        let req = Request::builder()
-            .uri("/api/schema_state/not-a-real-schema-id")
-            .method("GET")
-            .header("Host", "localhost")
-            .header("Authorization", admin_header)
-            .body(Body::empty())?;
-        let (parts, body) = partitioned_app
-            .router()
-            .clone()
-            .oneshot(req)
-            .await?
-            .into_parts();
-        let bytes = body.collect().await?.to_bytes();
-        let body = String::from_utf8_lossy(&bytes);
-        assert_eq!(parts.status, StatusCode::SERVICE_UNAVAILABLE, "{body}");
-        assert!(body.contains("ServiceUnavailable"), "{body}");
 
         Ok(())
     }
@@ -1310,7 +1256,7 @@ mod tests {
     }
 
     #[convex_macro::prod_rt_test]
-    async fn test_deploy2_routes_use_explicit_forwarding_authority(
+    async fn test_deploy_routes_use_explicit_forwarding_authority(
         rt: ProdRuntime,
     ) -> anyhow::Result<()> {
         let backend = setup_backend_for_test(rt).await?;
@@ -1325,7 +1271,7 @@ mod tests {
         partitioned_state.partition_id = Some(PartitionId(1));
         let partitioned_app = ConvexHttpService::new(
             router(partitioned_state),
-            "deploy2_authority_exemption_test",
+            "deploy_authority_exemption_test",
             SERVER_VERSION_STR.to_string(),
             MAX_CONCURRENT_REQUESTS,
             Duration::from_secs(125),
@@ -1333,7 +1279,7 @@ mod tests {
         );
 
         let req = Request::builder()
-            .uri("/api/deploy2/report_push_completed")
+            .uri("/api/deploy/report_push_completed")
             .method("POST")
             .header("Host", "localhost")
             .header("Content-Type", "application/json")
