+++ crates/local_backend/src/schema.rs
@@ -1,97 +1,28 @@
-use anyhow::Context;
-use application::deploy_config::ModuleJson;
-use axum::{
-    debug_handler,
-    extract::{
-        FromRequestParts,
-        State,
+use common::bootstrap_model::index::{
+    database_index::{
+        DatabaseIndexSpec,
+        DatabaseIndexState,
     },
-    response::IntoResponse,
-    RequestPartsExt,
-};
-use common::{
-    bootstrap_model::{
-        index::{
-            database_index::{
-                DatabaseIndexSpec,
-                DatabaseIndexState,
-            },
-            text_index::{
-                TextIndexSpec,
-                TextIndexState,
-            },
-            vector_index::{
-                VectorIndexSpec,
-                VectorIndexState,
-            },
-            IndexConfig,
-            IndexMetadata,
-        },
-        schema::{
-            invalid_schema_id,
-            parse_schema_id,
-            SchemaMetadata,
-            SchemaState,
-        },
+    text_index::{
+        TextIndexSpec,
+        TextIndexState,
     },
-    http::{
-        extract::{
-            Json,
-            MtState,
-            Path,
-        },
-        HttpResponseError,
+    vector_index::{
+        VectorIndexSpec,
+        VectorIndexState,
     },
-    runtime::Runtime,
-    types::IndexDiff,
+    IndexConfig,
+    IndexMetadata,
 };
-use database::{
-    IndexModel,
-    SchemaModel,
-};
-use errors::ErrorMetadata;
-use serde::{
-    Deserialize,
-    Serialize,
-};
+use serde::Serialize;
 use serde_json::{
     json,
     Value as JsonValue,
 };
 use value::{
     ConvexValue,
-    ResolvedDocumentId,
     TableName,
-    TableNamespace,
 };
-
-use crate::{
-    admin::{
-        must_be_admin,
-        must_be_admin_from_key_with_write_access,
-    },
-    authentication::ExtractAuthenticationToken,
-    DeployRouterState,
-};
-
-pub struct ExtractDeployIdentity(pub keybroker::Identity);
-
-impl FromRequestParts<DeployRouterState> for ExtractDeployIdentity {
-    type Rejection = HttpResponseError;
-
-    async fn from_request_parts(
-        parts: &mut axum::http::request::Parts,
-        st: &DeployRouterState,
-    ) -> Result<Self, Self::Rejection> {
-        let token: sync_types::AuthenticationToken =
-            parts.extract::<ExtractAuthenticationToken>().await?.into();
-        Ok(Self(
-            st.application
-                .authenticate(token, st.application.runtime().system_time())
-                .await?,
-        ))
-    }
-}
 
 #[derive(Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -210,196 +141,4 @@ impl TryFrom<IndexMetadata<TableName>> for IndexMetadataResponse {
             },
         })
     }
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PrepareSchemaArgs {
-    bundle: ModuleJson,
-    pub admin_key: String,
-    dry_run: Option<bool>,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PrepareSchemaResponse {
-    added: Vec<IndexMetadataResponse>,
-    dropped: Vec<IndexMetadataResponse>,
-    enabled: Vec<IndexMetadataResponse>,
-    disabled: Vec<IndexMetadataResponse>,
-    schema_id: String,
-}
-
-impl PrepareSchemaResponse {
-    fn new(diff: IndexDiff, schema_id: ResolvedDocumentId) -> anyhow::Result<Self> {
-        Ok(PrepareSchemaResponse {
-            added: diff
-                .added
-                .into_iter()
-                .map(IndexMetadataResponse::try_from)
-                .try_collect()?,
-            dropped: diff
-                .dropped
-                .into_iter()
-                .map(|doc| doc.into_value())
-                .map(IndexMetadataResponse::try_from)
-                .try_collect()?,
-            enabled: diff
-                .enabled
-                .into_iter()
-                .map(|doc| doc.into_value())
-                .map(IndexMetadataResponse::try_from)
-                .try_collect()?,
-            disabled: diff
-                .disabled
-                .into_iter()
-                .map(|doc| doc.into_value())
-                .map(IndexMetadataResponse::try_from)
-                .try_collect()?,
-            schema_id: schema_id.to_string(),
-        })
-    }
-}
-
-#[debug_handler]
-pub async fn prepare_schema(
-    State(st): State<DeployRouterState>,
-    Json(req): Json<PrepareSchemaArgs>,
-) -> Result<Json<PrepareSchemaResponse>, HttpResponseError> {
-    let (response, _) = prepare_schema_handler(st, req).await?;
-    Ok(response)
-}
-
-pub async fn prepare_schema_handler(
-    st: DeployRouterState,
-    req: PrepareSchemaArgs,
-) -> Result<(Json<PrepareSchemaResponse>, bool), HttpResponseError> {
-    let bundle = req.bundle.try_into()?;
-    let identity = must_be_admin_from_key_with_write_access(
-        st.application.app_auth(),
-        st.instance_name.clone(),
-        req.admin_key,
-    )
-    .await?;
-    let schema = match st.application.evaluate_schema(bundle).await {
-        Ok(m) => m,
-        Err(e) => return Err(e.into()),
-    };
-    let schema_validation_enabled = schema.schema_validation;
-    let mut tx = st.application.begin(identity.clone()).await?;
-
-    let dry_run = req.dry_run.unwrap_or(true);
-
-    // Table namespace is root because this endpoint is only used in non-components
-    // push.
-    let table_namespace = TableNamespace::root_component();
-    // In dry_run we only commit the schema, to enable CLI to check if the schema is
-    // valid.
-    let index_diff = if dry_run {
-        let mut tx = st.application.begin(identity.clone()).await?;
-        IndexModel::new(&mut tx)
-            .prepare_new_and_mutated_indexes(table_namespace, &schema)
-            .await?
-    } else {
-        IndexModel::new(&mut tx)
-            .prepare_new_and_mutated_indexes(table_namespace, &schema)
-            .await?
-    };
-
-    let (schema_id, schema_state) = SchemaModel::new(&mut tx, table_namespace)
-        .submit_pending(schema)
-        .await?;
-    let should_save_new_schema = match schema_state {
-        SchemaState::Pending => anyhow::Ok(true),
-        SchemaState::Validated | SchemaState::Active => Ok(false),
-        SchemaState::Failed { .. } | SchemaState::Overwritten => Err(anyhow::anyhow!(
-            "Newly inserted pending schema cannot be failed or overwritten."
-        )),
-    }?;
-
-    if index_diff.is_empty() && !should_save_new_schema {
-        drop(tx);
-    } else {
-        let audit_events = if !dry_run && !index_diff.is_empty() {
-            vec![index_diff.clone().into()]
-        } else {
-            vec![]
-        };
-        st.application
-            .commit_with_audit_log_events(tx, audit_events, "prepare_schema")
-            .await?;
-    }
-
-    Ok((
-        Json(PrepareSchemaResponse::new(index_diff, schema_id)?),
-        schema_validation_enabled,
-    ))
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct SchemaStateResponse {
-    indexes: Vec<IndexMetadataResponse>,
-    schema_state: SchemaStateJson,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-#[serde(tag = "state")]
-enum SchemaStateJson {
-    Pending,
-    Validated,
-    Active,
-    #[serde(rename_all = "camelCase")]
-    Failed {
-        error: String,
-        table_name: Option<String>,
-    },
-    Overwritten,
-}
-
-impl From<SchemaState> for SchemaStateJson {
-    fn from(state: SchemaState) -> Self {
-        match state {
-            SchemaState::Pending => SchemaStateJson::Pending,
-            SchemaState::Validated => SchemaStateJson::Validated,
-            SchemaState::Active => SchemaStateJson::Active,
-            SchemaState::Failed { error, table_name } => {
-                SchemaStateJson::Failed { error, table_name }
-            },
-            SchemaState::Overwritten => SchemaStateJson::Overwritten,
-        }
-    }
-}
-
-/// Gets the current state of the indexes and schema.
-pub async fn schema_state(
-    MtState(st): MtState<DeployRouterState>,
-    Path(schema_id): Path<String>,
-    ExtractDeployIdentity(identity): ExtractDeployIdentity,
-) -> Result<impl IntoResponse, HttpResponseError> {
-    must_be_admin(&identity)?;
-    let mut tx = st.application.begin(identity.clone()).await?;
-    // This endpoint is only used in non-components push.
-    let table_namespace = TableNamespace::root_component();
-    let indexes = IndexModel::new(&mut tx)
-        .get_application_indexes(table_namespace)
-        .await?;
-    let schema_id = parse_schema_id(&schema_id, tx.table_mapping(), table_namespace)
-        .context(invalid_schema_id(&schema_id))?;
-
-    let doc = tx.get(schema_id).await?.ok_or_else(|| {
-        anyhow::anyhow!(ErrorMetadata::not_found(
-            "SchemaNotFound",
-            format!("Schema with id {schema_id} not found"),
-        ))
-    })?;
-    let SchemaMetadata { state, .. } = doc.into_value().into_value().try_into()?;
-    Ok(Json(SchemaStateResponse {
-        indexes: indexes
-            .into_iter()
-            .map(|idx| idx.into_value().try_into())
-            .collect::<anyhow::Result<_>>()?,
-        schema_state: state.into(),
-    }))
 }
