+++ npm-packages/system-udfs/convex/_system/frontend/deploymentEvents.ts
@@ -2,19 +2,25 @@ import { Doc } from "../../_generated/dataModel";
 import { queryPrivateSystem } from "../secretSystemTables";
 import { v } from "convex/values";
 
-type PushConfigEvent = Doc<"_deployment_audit_log"> & {
-  action: "push_config" | "push_config_with_components";
+type DeployEvent = Doc<"_deployment_audit_log"> & {
+  action:
+    | "deploy"
+    | "deploy_with_components"
+    | "push_config"
+    | "push_config_with_components";
 };
 
-export const lastPushEvent = queryPrivateSystem({
+export const lastDeployEvent = queryPrivateSystem({
   args: {
     componentId: v.optional(v.union(v.string(), v.null())),
   },
-  handler: async function ({ db }): Promise<PushConfigEvent | null> {
-    const lastPushEvent = await db
+  handler: async function ({ db }): Promise<DeployEvent | null> {
+    const lastDeployEvent = await db
       .query("_deployment_audit_log")
       .filter((q) =>
         q.or(
+          q.eq(q.field("action"), "deploy"),
+          q.eq(q.field("action"), "deploy_with_components"),
           q.eq(q.field("action"), "push_config"),
           q.eq(q.field("action"), "push_config_with_components"),
         ),
@@ -25,13 +31,15 @@ export const lastPushEvent = queryPrivateSystem({
     // Making the type system happy. This should never happen
     // because the query should always return a push event.
     if (
-      lastPushEvent &&
-      lastPushEvent.action !== "push_config" &&
-      lastPushEvent.action !== "push_config_with_components"
+      lastDeployEvent &&
+      lastDeployEvent.action !== "deploy" &&
+      lastDeployEvent.action !== "deploy_with_components" &&
+      lastDeployEvent.action !== "push_config" &&
+      lastDeployEvent.action !== "push_config_with_components"
     ) {
       return null;
     }
 
-    return lastPushEvent;
+    return lastDeployEvent;
   },
 });
